split user component in admin component
